Remove Client.ts (#77)
## Description

This pull request removes the mediator Client class and moves the
functionality into smaller React Hooks.
Additionally, we decided to remove the `useSetupMedia` hook.

## Motivation and Context

The main motivation is to remove the Client class from react client. We
only need to synchronize the state with the TS Client. All additional
state can be already maintained in the react hooks.

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to
      not work as expected)