FCE 410 Add toggle function (#95)
## Description

The toggle function works in two modes: "soft" and "hard" (default). The
user can invoke this function without any arguments. It starts (or
resumes) the device and streams media if the user is connected.

### `hard` 
- Turns off the device
- Slower start
- Turns off the device LED
- Example usage: camera

### `soft`
- Only disables the device
- Faster start
- The device LED remains active
- Example usage: microphone

## Motivation and Context

To unify the API with the Mobile SDK. It's important to implement this
feature to avoid confusion:

https://linear.app/swmansion/issue/FCE-489/add-the-camera-and-mic-after-connecting-if-they-are-active-web

## Types of changes

New feature (non-breaking change which adds functionality)