FCE-759 Hide the error if the user closes the screen share window (#157)
## Description

When the screen share window is closed, `NotAllowedError` is thrown. The
`Share the screen` handler swallows this error.

## Motivation and Context

When the user changes their mind during screen sharing and closes the
browser window, an error appears in the console. This is a completely
normal situation, so we should not log this error to the console.

## Types of changes

Bug fix (non-breaking change which fixes an issue)

Co-authored-by: Adrian Czerwiec <adrian.czerwiec@swmansion.com>