FCE 756 enable unit test and remove validators (#165)
## Description

- Generics were removed from `webRTCEndpoint.ts`.
- Metadata parsers were removed from `webRTCEndpoint.ts`.
- The `rawMetadata` and `metadataParsingError` fields were removed.
- Some unit tests were fixed.
- Unit and E2E tests that tested metadata parsers were deleted.
- Fishjam in E2E tests was updated to version 0.9.0.
- Unit tests were enabled to run on every PR.



## Motivation and Context
**Why were metadata parsers removed from `webRTCEndpoint.ts`?**

Because Fishjam changed the peer metadata structure. Fishjam introduced
a separation between peer and server metadata. Previously, if a peer set
its metadata to `{ a: "sth" }`, Fishjam would broadcast this change to
other peers as the same object (`{ a: "sth" }`). Now, if the user sets
the same object, Fishjam broadcasts it as `{ peer: { a: "sth" }, server:
{ b: "other server metadata" } }`. I didn't want to introduce more
complicated parsers, so I deleted them.

## Types of changes

Breaking change (fix or feature that would cause existing functionality
to not work as expected)

---------

Co-authored-by: Adrian Czerwiec <adrian.czerwiec@swmansion.com>
Co-authored-by: Adrian Czerwiec <33912477+czerwiukk@users.noreply.github.com>